requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
jinja2==3.1.2
base64io==1.0.3
colorthief==0.2.1
